Export ZIPs when the app is updated
It was brought to my attention that when I pushed the migration update to
the WireGuard-KT, the undimissible dialog caused many users to be
locked out of the app with no way to export their configs, making for a
very disastrous blunder on my end causing a rather bad UX for users.

To avoid something like it happening again, install an update receiver
that exports all configurations each time the app is updated so broken
updates can be rolled back without trouble. I hope nobody has to rely on
this ever, but I'm only human and things can always slip by. This does
not fire on debugging builds since I install way too often and the IO
penalty is not worth it.

Signed-off-by: Harsh Shandilya <msfjarvis@gmail.com>