AddTunnelsSheet: Fix ZIP/conf import
When the options were moved out from their
previous home in TunnelListFragment, all startActivityForResult
calls would notify the AddTunnelsSheet DialogFragment which served
no purpose, because we never handle import requests there. To work
around the requirement of needing to pass the activity result to
TunnelListFragment, make the bottom sheet class accept a fragment
parameter which we then use to call startActivityForResult with,
resulting in the right fragment being notified of this activity result.

I had to get out of bed for this so everybody better be grateful for the
fix and sorry for not spotting it themselves.

Signed-off-by: Harsh Shandilya <msfjarvis@gmail.com>