Introduce TunnelToggleActivity
On Android 10, apps cannot start services when they're in the background. This means that
starting VpnService from within QuickTileService when the app is not active ends badly. To
mitigate this situation, we introduce a proxy activity of sorts that will handle starting
VpnService for us. The activity is completely transparent and invisible, and does only four
things:

- Toggle the tunnel state
- Request the Tile bound by QuickTileService to refresh its state
- Handle any error that might have been thrown during toggle
- Call finish() and go away

Since this service-activity-service cycle is a bit slow, we restrict to using this only when
we absolutely need to: On Android 10, using GoBackend, when starting the tunnel.

Signed-off-by: Harsh Shandilya <msfjarvis@gmail.com>