TunnelListFragment: Fix occasional crash with search/detail view
savedTunnelList holds a list of all active tunnels but fails to properly keep them in sync with the adapter.

When this fail happens, we'll see ObservableSortedKeyedArrayList throw an exception that keys with the same name already exist.

To reproduce:

1. Import multiple tunnels from a ZIP
2. Delete all of them
3. Import them again
4. Try to search or open the detail fragment
5. App crashes with the following stacktrace

java.lang.IllegalArgumentException: Element with same key already exists in list
    at com.wireguard.android.util.ObservableSortedKeyedArrayList.add(ObservableSortedKeyedArrayList.java:39)
    at com.wireguard.android.util.ObservableSortedKeyedArrayList.addAll(ObservableSortedKeyedArrayList.java:58)
    at com.wireguard.android.fragment.TunnelListFragment.onQueryTextChange(TunnelListFragment.kt:318)

Signed-off-by: Harsh Shandilya <msfjarvis@gmail.com>